fix(lib/realm): fix MISRA C:2012 Rule 10.3 in status.h

+++ lib/realm/include/status.h
@@ -61,7 +61,8 @@ static inline unsigned long pack_return_code(unsigned int status, unsigned int i
  */
 static inline return_code_t unpack_return_code(unsigned long error_code)
 {
-	return make_return_code(error_code & 0xffU, error_code >> 8);
+	return make_return_code((unsigned int)error_code & 0xffU,
+				(unsigned int)error_code >> 8);
 }
 
 #define MAX_ERR 4095
